Safari-611.1.10/LayoutTests/http/wpt/beacon/beacon-async-error-logging-expected.txt

CONSOLE MESSAGE: Beacon API cannot load http://localhost:1/. Not allowed to use restricted network port

PASS Should log an error message in the console
